Moved StaleDataCheckStep to spark/delta.py + minor improvements
